agents/requirements.txt
smolagents
anthropic
requests
geopy
networkx
uvicorn
fastapi
redis
asyncpg
black
isort
pytest
python-dotenv
pydantic>=2.0
